BACKEND: Modularización del Sistema Selnic, antes llamado DMS
