+++ backend/src/modules/administration/models/Departamento.model.js
@@ -1,5 +1,4 @@
-import { DataTypes } from "sequelize";
-import { sequelize } from "../config/db.config.js";
+import { sequelize } from "../../../config/db.config.js";
 
 export const Departamento = sequelize.define(
   "departamento",
