+++ backend/src/modules/administration/models/IntentoFallido.model.js
@@ -1,5 +1,4 @@
-import { DataTypes } from "sequelize";
-import { sequelize } from "../config/db.config.js";
+import { sequelize } from "../../../config/db.config.js";
 import { Usuario } from "./Usuario.model.js";
 
 export const IntentoFallido = sequelize.define("intento_fallido", {
