+++ backend/src/modules/administration/models/SystemParameter.model.js
@@ -1,5 +1,4 @@
-import { DataTypes } from "sequelize";
-import { sequelize } from "../config/db.config.js";
+import { sequelize } from "../../../config/db.config.js";
 
 export const ParametroSistema = sequelize.define(
   "sis_parametros",
