+++ backend/src/modules/document-management/models/TramiteAsignacion.model.js
