+++ backend/src/modules/document-management/models/TramiteEliminacion.model.js
