--- backend/src/services/auth.service.js
@@ -1,1 +1,0 @@
-// Servicio para manejo de autenticación (login, tokens, etc.)
