--- backend/src/services/fileUpload.service.js
@@ -1,1 +1,0 @@
-// Servicio para subir y gestionar archivos (Multer)
